+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Authentications Log (Thanks, [@weining-li](https://github.com/weining-li))
 - Brand Configs (Thanks, [@bennettscience](https://github.com/bennettscience))
 - Comm Messages (Thanks, [@ljoks](https://github.com/ljoks))
+- File Usage (Thanks, [@atarisafari](https://github.com/atarisafari))
 - Quiz Submission Events (Thanks, [@Goff-Davis](https://github.com/Goff-Davis))
 - Rubric Associations (Thanks, [@weining-li](https://github.com/weining-li))